Recognition should not be importing net globally.
Otherwise we can't exclude tf & keras from the prod build
