+++ digito-api/config.py
@@ -1,0 +1,4 @@
+
+MNIST_ROWS = 28
+MNIST_COLS = 28
+PADDING = 3
